refactor: remove sheets dependency on engine-render
