+++ packages/engine-render/src/basics/tools.ts
@@ -18,13 +18,12 @@ import type {
     IRange,
     IRangeWithCoord,
     IScale,
-    ISelectionCell,
     ISelectionCellWithCoord,
     IStyleBase,
     LocaleService,
     Nullable,
 } from '@univerjs/core';
-import { BaselineOffset, ColorKit, DEFAULT_STYLES, FontStyleType, Rectangle, Tools } from '@univerjs/core';
+import { BaselineOffset, ColorKit, DEFAULT_STYLES, FontStyleType, getCellInfoInMergeData, Rectangle, Tools } from '@univerjs/core';
 import * as cjk from 'cjk-regex';
 
 import { FontCache } from '../components/docs/layout/shaping-engine/font-cache';
@@ -596,69 +595,6 @@ export function getCellByIndex(
         startX,
         endX,
         mergeInfo,
-    };
-}
-
-/**
- * Determines whether the cell(row, column) is within the range of the merged cells.
- */
-export function getCellInfoInMergeData(row: number, column: number, mergeData?: IRange[]): ISelectionCell {
-    let isMerged = false; // The upper left cell only renders the content
-    let isMergedMainCell = false;
-    let newEndRow = row;
-    let newEndColumn = column;
-    let mergeRow = row;
-    let mergeColumn = column;
-
-    if (mergeData == null) {
-        return {
-            actualRow: row,
-            actualColumn: column,
-            isMergedMainCell,
-            isMerged,
-            endRow: newEndRow,
-            endColumn: newEndColumn,
-            startRow: mergeRow,
-            startColumn: mergeColumn,
-        };
-    }
-
-    for (let i = 0; i < mergeData.length; i++) {
-        const {
-            startRow: startRowMarge,
-            endRow: endRowMarge,
-            startColumn: startColumnMarge,
-            endColumn: endColumnMarge,
-        } = mergeData[i];
-        if (row === startRowMarge && column === startColumnMarge) {
-            newEndRow = endRowMarge;
-            newEndColumn = endColumnMarge;
-            mergeRow = startRowMarge;
-            mergeColumn = startColumnMarge;
-
-            isMergedMainCell = true;
-            break;
-        }
-        if (row >= startRowMarge && row <= endRowMarge && column >= startColumnMarge && column <= endColumnMarge) {
-            newEndRow = endRowMarge;
-            newEndColumn = endColumnMarge;
-            mergeRow = startRowMarge;
-            mergeColumn = startColumnMarge;
-
-            isMerged = true;
-            break;
-        }
-    }
-
-    return {
-        actualRow: row,
-        actualColumn: column,
-        isMergedMainCell,
-        isMerged,
-        endRow: newEndRow,
-        endColumn: newEndColumn,
-        startRow: mergeRow,
-        startColumn: mergeColumn,
     };
 }
