+++ packages/sheets/package.json
@@ -63,7 +63,6 @@
     "peerDependencies": {
         "@univerjs/core": "workspace:*",
         "@univerjs/engine-formula": "workspace:*",
-        "@univerjs/engine-render": "workspace:*",
         "@univerjs/rpc": "workspace:*",
         "@wendellhu/redi": "0.15.4",
         "rxjs": ">=7.0.0"
@@ -75,7 +74,6 @@
         "@univerjs/core": "workspace:*",
         "@univerjs/engine-formula": "workspace:*",
         "@univerjs/engine-numfmt": "workspace:*",
-        "@univerjs/engine-render": "workspace:*",
         "@univerjs/rpc": "workspace:*",
         "@univerjs/shared": "workspace:*",
         "@wendellhu/redi": "0.15.4",
