+++ packages/sheets/src/commands/mutations/set-worksheet-row-height.mutation.ts
@@ -14,9 +14,8 @@
  * limitations under the License.
  */
 
-import type { BooleanNumber, IMutation, IObjectArrayPrimitiveType, IRange, Nullable, Workbook } from '@univerjs/core';
+import type { BooleanNumber, IMutation, IObjectArrayPrimitiveType, IRange, IRowAutoHeightInfo, Nullable, Workbook } from '@univerjs/core';
 import { CommandType, IUniverInstanceService } from '@univerjs/core';
-import type { IRowAutoHeightInfo } from '@univerjs/engine-render';
 import type { IAccessor } from '@wendellhu/redi';
 
 const MAXIMUM_ROW_HEIGHT = 2000;
